third_party/blink/web_tests/external/wpt/css/selectors/focus-visible-024-expected.txt

This is a testharness.js-based test.
PASS ":focus-visible" should be a valid selector
FAIL :focus-visible matches after accesskey on DIV assert_equals: outline-style for DIV#targetA should be solid expected "solid" but got "none"
FAIL :focus-visible matches after accesskey on BUTTON assert_equals: outline-style for BUTTON#targetB should be solid expected "solid" but got "none"
Harness: the test ran to completion.
